Allow to pass undefined for createMutation/deleteMutation/updateMutation in getMutationsPromises function

+++ assembl/static2/js/app/components/administration/saveButton.jsx
@@ -15,21 +15,21 @@ export const getMutationsPromises = (params) => {
   const { items, variablesCreator, deleteVariablesCreator, createMutation, deleteMutation, updateMutation, lang } = params;
   const promises = [];
   items.forEach((item, index) => {
-    if (item.isNew && !item.toDelete) {
+    if (item.isNew && !item.toDelete && createMutation) {
       // create item
       const payload = {
         variables: variablesCreator(item, index)
       };
       const p1 = () => createMutation(payload);
       promises.push(p1);
-    } else if (item.toDelete && !item.isNew) {
+    } else if (item.toDelete && !item.isNew && deleteMutation) {
       // delete item
       const payload = {
         variables: deleteVariablesCreator(item)
       };
       const p3 = () => deleteMutation(payload);
       promises.push(p3);
-    } else {
+    } else if (updateMutation) {
       // update item
       const variables = variablesCreator(item, index);
       variables.id = item.id;
